backend/pyproject.toml
[project]
name = "myfuture-ai-backend"
version = "0.1.0"
description = "MyFuture AI backend (FastAPI)"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic-settings>=2.2",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "passlib[bcrypt]>=1.7",
  "python-jose[cryptography]>=3.3",
  "httpx>=0.27",
]

[tool.uvicorn]
factory = false

[tool.ruff]
line-length = 100

[tool.pyright]
venvPath = "."
